httpreplay: part 1: proxy for recording
This is the first CL of an HTTP proxy that will allow recording and replaying of
traffic from the HTTP-based Google API clients. The proxy will be available via both
an in-memory API and as a standalone binary.

This CL contains half of the package that will be shared between the
API and the binary. It uses the Martian proxy package to set up a
proxy, and adds modifiers to it to log requests and responses in the
standard HAR format. It also contains code to redact or avoid logging
auth secrets like OAuth2 tokens.

Change-Id: I63a3e1e882c8c679cbb1353337fab1b57fec7150
Reviewed-on: https://code-review.googlesource.com/27290
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>