logging: change trace format
When setting the trace ID in a log entry, we were using the value of the X-Cloud-Trace-Context header
directly. According to https://cloud.google.com/appengine/docs/flexible/go/writing-application-logs,
we should use "projects/{P}/traces/{T}", where T is the header value.

This CL makes that change, at least for logging clients whose parent is a project. If the parent
is a folder, org or billing account, we'll use the same parent as the prefix in the trace string.
It's not clear whether this is right, but it's the best we can do.

Change-Id: Id97c799ae0f7706bfc15cfd23a07b7c3112b9581
Reviewed-on: https://code-review.googlesource.com/31390
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>