pubsub: send initial modack in synchronous=true mode
Messages that do not have an initial modack sent often become invisibly
expired. How this happens:

(given some ack distribution spooled up to 5m)
(given subscription config AckDeadline=15s)

- kaTick happens, send modacks. Next kaTick is 5m later
- 10s after this kaTick occurs, a message arrives. We do not send a modack, so
the expiration is 15s.
- 15s goes by. If the user didn't ack by now this message is dead-in-the-water,
because now the server has decided that this message has expired.
- 4m35s goes by and we send a modack. But, of course, it has already expired.

Whenever the user gets around to sending the ack, if they do so after the 15s
this ack will be recorded as status=expired in stackdriver.

This CL makes all messages - including synchronous - send a modack on receipt
to fix this issue. Users relying on Synchronous=true as a way to force the
subscription's AckDeadline to be used should instead use MaxExtension:

```
	cfg, err := sub.Config(ctx)
	if err != nil {
		// TODO handle err
	}

	sub.ReceiveSettings.MaxExtension = cfg.AckDeadline
```

This CL also removes the customized logger in favour of t.Log and a panic.

Fixes #1247

Change-Id: I6aa65e083ab12a79e61e8a4d968910dc794c964d
Reviewed-on: https://code-review.googlesource.com/c/36430
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Eno Compton <enocom@google.com>