httpreplay: remove and redact instead of ignore
This is the first in a series of CLs that makes the httpreplay
package more configurable and powerful.

1. Instead of ignoring some headers at replay, remove them from
the log when recording, and the incoming request when replaying.
This saves on log space when we don't care about the value, or don't
want it to show (like secrets).

2. Support redacting but keeping a header, when we want to hide a
secret but we still want to check that the header is present.

3. Add a converter type that does all this. Use it both when we log
and when we replay, then compare the resulting Requests.

Change-Id: Ic7e009396ace4d28cdc843dbdada876fbeb87222
Reviewed-on: https://code-review.googlesource.com/c/37210
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Eno Compton <enocom@google.com>