datastore: add error checking to Query.nextBatch
Currently, nextBatch will panic if the given query is somehow invalid.
This can happen when Filter is invoked with an invalid type:

```
type MyType int
var v MyType = 1
q := datastore.NewQuery("MyKind").Filter("fld=", v)
n, err := c.Count(ctx, q)
```

This happens because the type validation happens while running the query
and converting the query to protobuf (via toProto). If this fails, the
iterator will set an error, however nextBatch is not checking the error.
Therefore the iterator is in an invalid state and certain fields are
nil. When the user invokes Next again, the call panics (in nextBatch).

This CL fixes this problem by having nextBatch check the iterator's err
field.

Fixes #1306

Change-Id: I8cfe48a17258a1ab4a18580230bffc51e3b2c75c
Reviewed-on: https://code-review.googlesource.com/c/38170
Reviewed-by: Jean de Klerk <deklerk@google.com>