bigquery: introduce models API.
This change exposes the BigQuery models API.  It allows for
get, list, update, delete operations, but does not allow
direct model creation via the API, as a CREATE MODEL query
remains the single way in which a new ML model can be created.

Prior to this, models were exposed as a type of Table.

Models support limited mutation, which can be done through the
use of ModelMetadataToUpdate.

Much of the value of a model is in the stats.  However, it is
unclear how this should be presented absent community feedback.
As an interim measure, this exposes three Raw* methods on
ModelMetadata to expose information about label columns, feature
columns, and training runs which leverages type representations
from google.golang.org/api/bigquery/v2.

Change-Id: I68e94410dd0e14260ba8b2e03befde8324e6e96c
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/39090
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>