bigtable, spanner: consolidate dupe OpenCensus tracing code
Packages bigtable and spanner had identical code in their
trace.go files for starting and ending spans. On ending
spans, if we encountered an error usually from a remote RPC,
each traceEndSpan variant function would just set the span's
Status.Message as the error but unfortunately would keep the
Status.Code as 0 which defaults to Status.OK which is not correct,
and can be misdiagnosed during trace examination.

This CL delegates this functionality to internal/trace which already has
the correct code. That package has comprehensive code that correctly
extracts statuses and messages and converts to them be settable on
the created spans.

Added a TODO to later on perhaps keep the result of:
   ctx, span := trace.StartSpan(ctx, spanName)
instead of passing around only the context and having
to incur extra costs from trace.FromContext(ctx) on every call.
However, we'll have to evaluate these costs if they are actually
expensive and a priority.

Also noticed that perhaps in a later CL we should rename internal/trace
to internal/observability, because some of these packages uses stats
as well.

Fixes #1376

Change-Id: I22da12ad23d4885322636c18e3dd5f2897a9f1a5
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/39410
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>