httpreplay: set up MITM config only once
Creating a certificate and key pair for the proxy's man-in-the-middle
support took some time (.5 s). This really slowed down repeated proxy
creation, as in a record/replay test where each test function is
recorded to a separate file.

Instead, create the MITM config once per process, and reuse it for
all proxies.

Change-Id: Ied28735352552813227dacca6b20fab5c05478b6
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/40510
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Eli Bendersky <eliben@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>