pubsub: add ExpirationPolicy to Subscription
Adds ExpirationPolicy a field that specifies the conditions
for a subscription's expiration.

If ExpirationPolicy is not set, Pubsub's backend uses the
default of 31 days. The minimum allowed value for ExpirationPolicy
is 1 day.

This CL is a re-revert of CL
   https://code-review.googlesource.com/c/gocloud/+/40931
which was failing just because now we deserialize the default value
sent back by the server of 31 days aka 744hr0m.
This change updates the previously failing integration test to expect
the deserialized and default value sent back by Pubsub.

This change also updates an example to show how to use
ExpirationPolicy.

Updates #1361

Change-Id: Ifd44c3aa52923d42f060b9328e78a01878a94084
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/40972
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>