pubsub: fix TestPublishTimeout
Not sure how this was working before. Neither internal/testutil/ or pubsub/
has had any changes related to this, but the fix seems obvious: the server
never responds because the server never starts.

The difference between the server responding with Unavailable and the server
never responding when in the presence of PublishSettings.Timeout is that the
latter results in a context.Context deadline exceeded error, whereas the latter
results in an RPC deadline exceeded error. The test wants the former.

Change-Id: I4bd0ce5d2cf26dc3a439211681c7deeeae17f2ad
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/41050
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Emmanuel Odeke <emmanuel@orijtech.com>