unflake TestHealthCheckScheduler
The TestHealthCheckScheduler checked the number of pings per session that
it created. It did however not take into account that the session pool
might delete one of those sessions while the test is running. The number
of pings on that session would then never reach the minimum number of
pings needed.

Fixes #1491

Change-Id: I7ea142162389d76896ec013fda49fa14dfdc5604
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/42753
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>