logging: add auto-detection of monitored resources on GAE Standard
This package can auto-detect monitored resources on GCE.  This change
adds detection of monitored resources if running on GAE Standard (second
generation) too.

I got the label-names we need from [1] and the envvar names from [2]
(which is written for Go 1.12, but I believe applies to all second-gen
runtimes).

I wasn't sure if there was a useful way to write automated tests, but I
put similar code in my app (via the `CommonResource` option), and
checked that it caused my log lines to show up in the logs viewer under
"GAE Application" (and/or the relevant module/version) rather than
"Google Project".

[1] https://cloud.google.com/monitoring/api/resources#tag_gae_app
[2] https://cloud.google.com/appengine/docs/standard/go112/runtime#environment_variables

Fixes #1521.

Change-Id: Ie7243c16fb5b10ac06a57bdcb526c247b0e215be
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/42810
Reviewed-by: Steven Buss <sbuss@google.com>
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Emmanuel Odeke <emm.odeke@gmail.com>