storage: remove duplicated retry logic
Removes the duplicated retry logic that operated on the premise
that since the logic from googleapis/gensupport.Resumable didn't retry
the initial HTTP request that created a resumable upload session,
so we'd retry it. However, for other HTTP requests such as when
uploading chunks, gensupport.Resumable actually retried them, so
really we ended up retrying twice the entirety of exponential backoff.

On the event of a 429 or 5XX error, the logic from
googleapis/gensupport.Resumable would run its course of retries e.g.
    [0, 2**n]
and then fallback to the storage package's fallback which would then
re-invoke a retry and also run the course of retries of a magnitude:
    [0, 2**m]

which basically meant that we'd retry altogether:
    (2**n) * (2**m) = (2**(n+m))
times.
This could easily become a very large number
say if n=10, m=10 and say the base unit was a second:
    (2**20)seconds = ~12.13 days

This change adds a regression test too to ensure
that we don't regress from indefinite retries.

There is still a lurking inconsistency from googleapis/gensupport in
https://github.com/googleapis/google-api-go-client/issues/392, but
for now it is better having the initial request not being retried
than with indefinite retries but also when that issue fixed, all
we'll need to do is have a single version update in our go.mod
file.

Fixes #1507
Follow-up https://github.com/googleapis/google-api-go-client/issues/392

Change-Id: I676c06d07194ba0522194bba72a6386ecbc62391
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/44071
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>