storage/internal/benchserver: make body reads idempotent
When we converted "files" to be
    map[string]io.Reader
we used an io.LimitedReader that then unfortunately
meant that multiple runs wouldn't stream back the
content. That's because when io.LimitedReader hits
io.EOF, we need to reset its N for it to read
more from rand.Reader.

This would then mean that the second request would
stream back 0 bytes, which is not what we want.

An update to make this reset would be a bit complex,
so instead make "files"
    map[string]int64
mapping the file name to its size. When a request
is then made for the file, we'll create a fresh
io.LimitedReader, which produces the correct behavior.

Change-Id: I609da9458fc28efece4967a73520b5ac6214f6cb
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/44191
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Alex Hong <hongalex@google.com>