storage: add HMACKeyOptions
Added options to give users control over
various HMACKey methods like for:

a) Listing
we now provide three filtering options:
    storage.ShowDeletedHMACKeys()
    storage.ForHMACKeyServiceAccountEmail(<SERVICE_ACCOUNT_EMAIL>)
    storage.HMACKeysForUserProject(<USER_PROJECT_ID>)

* storage.HMACKeysForUserProject is currently ignored by the
storage API but we are including it here because at some
so that if we forget to update the source code, whenever it
the storage team and APIs enable it, it won't require CLs
which are harder than just removing a source code comment
and an announcment.

The options are drawn from the document GCS HMAC Key User guide
and PRD as well as from the official storage docs:
https://cloud.google.com/storage/docs/json_api/v1/projects/hmacKeys/list

Updates #1486.

Change-Id: I0a8d0ba12506ff929855f6f4aa78dc760f697ed6
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/44390
Reviewed-by: Jean de Klerk <deklerk@google.com>
Reviewed-by: Frank Natividad <franknatividad@google.com>