bigquery: address issue with RowIterator behavior
Existing behavior for row iterators was to nil out the pagefetcher
when reading a table/result with zero rows.  While fine for the
normal iteration case, this causes issues when wrapping an iterator in a
pager.

This change drops the special case behavior, at the (potential) cost
of a unnecessary roundtrip for zero result queries.

This change also adds some simple literal testing to verify behavior,
since we do not currently do much testing for corner cases like zero
result queries.

Fixes: #1642

Change-Id: I919840af660903940fa0f60e6afd5e2d9cc39d19
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/47910
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>