spanner: Append release notes
The change of the default for SessionPoolConfig.WriteSessions to 0.2
was not included in the original release notes. That change would
break clients that were created using the default NewClient method
while using credentials that did not have permission to create
read/write transactions on the Spanner databases they were connected
to.

Updates #1665.

Change-Id: Id425ab59c2a8d10aa3b5b757763c85856a906251
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/48530
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hengfeng Li <hengfeng@google.com>
Reviewed-by: Shanika Kuruppu <skuruppu@google.com>