spanner: extract retry info from status
The Spanner client used to extract RetryInfo from the trailers of a
gRPC request. This meant that an extra option had to be added to the
gRPC call to ensure that the trailers were parsed. It also meant that
these trailers needed to be kept in a separate field in spanner.Error.

RetryInfo and other specific error details are however also included in
the wrapped statusError. Instead of getting this information from the
trailers of the request, the Spanner client should get it directly from
the wrapped statusError. This makes it less error prone, as we don't
have to specify extra options for the RPCs where we might want trailers
to be parsed. It also prepares the Spanner client for getting other
additional information from the wrapped statusError, such as additional
information on the type of resource that was not found. This will
allow us to stop parsing textual error messages to determine whether an
error was a 'Session not found' error, and instead use the details from
the statusError.

Fixes #1813.

Change-Id: I9fab63c5f2e3c8d632f136fe3822c170318c5d78
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/52790
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hengfeng Li <hengfeng@google.com>