storage: properly read files with Content-Encoding gzip
Dictated by Cloud Storage's "decompressive transcoding" advisory:
    https://cloud.google.com/storage/docs/transcoding

If a file is stored with
    "Content-Encoding": "gzip"
that file is stored on Cloud Storage in the compressed form but when
downloaded by clients it will be decompressed on the fly and
served in the decompressed form.

That also means that range requests for such files are ignored by Cloud Storage.

This change removes the prior assumption that EVERY range
request would result in a 206 Partial Content response,
yet in this case a whole file is always served regardless
of the range for files with "Content-Encoding": "gzip".
Instead such files are served back with a 200 status code.

Due to the consequence of files being served wholly by Cloud Storage
irregardless of the range request, we manually consume "seen" bytes from
retried bodies so that we only send back the caller the remaining bytes.

This change adds both integration and behavioral/mocking tests
to ensure that the behavior is always the same when dealing with
such files.

Also manually verified by testing out a repro that was created
for the issue and checksumming the response. The reproducer isn't
included in the tests because it requires the server to panic,
causing an HTTP/2 stream error, then perfom the retry. However,
the results of serving the retried file, after MD5 checksumming a file
containing 1 million lines of "I am a line\n":

    GotMD5:  90d6ab70c7e44d6b3ccdec8658b56f69
    WantMD5: 90d6ab70c7e44d6b3ccdec8658b56f69

Fixes #1800

Change-Id: I858e547e5faa5d2caa70b18b64f4d03dcaac8270
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/54791
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Chris Cotter <cjcotter@google.com>
Reviewed-by: Tyler Bui-Palsulich <tbp@google.com>
Reviewed-by: Frank Natividad <franknatividad@google.com>