internal/gapicgen: use draft PRs for google-cloud-go (#2475)
Now that we use CODEOWNERS to control assigning of reviewers we
have a little problem. When a PR is created it will auto-assign
reviewers based on where there is modified code. To get around this
fact, if we don't want reviewers assigned we can use draft PRs.

Sadly the functionality to switch from a draft PR to a reviewable
PR does not exist in the v3 REST API for GitHub. You need to use
v4 which is GraphQL. I had to add a dependency to add this
functionality. The project is maintained by a Googler, and I
confirmed everything works as expected in #2473.