refactor(all): refactor examples to reduce dependency tree (#2507)
Reworked examples to use a client in the top module as to no
longer incure extra direct dependencies. In the future we consider
refactoring the tests that use the storage gapic to instead use
the apiary to remove the last sub-module dependency.

After another release of gocloud, a release storage, and a
`go mod tidy` our `go.sum` files should start to look a lot
cleaner.