bigquery: refactor RowIterator to support fast query path (#2484)
This CL significantly revamps the rowiterator.

Previously, row iteration necessitated a reference to a Table object.
For jobs, this meant we would consult job statistics to find the
reference to the destination table.

With this change, we switch to using the jobs.getQueryResults RPC to get
rows from a job, while still allowing table reads to use the existing
tabledata.list mechanism.

This change also includes support for a caching abstraction in the row
iterator, which is added but not yet leveraged.  This will be wired up
in subsequent PRs that change the query execution path.

Additionally, this PR address some other reported issues against the library as the use of jobs.getQueryResults allows the service to make decisions about the source of results (e.g. in cases where a scripting/multistatement query returns resultsets from a child job).

Fixes: #1960, #1974